[Feat] #5 refactor ViewModel
