+++ app/src/main/AndroidManifest.xml
@@ -19,22 +19,16 @@
 
         <activity
             android:name=".login.LoginActivity"
-            android:exported="false">
-        </activity>
-
-        <activity
-            android:name=".signup.SignUpActivity"
-            android:exported="false">
-        </activity>
-
-        <activity
-            android:name=".home.HomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
 
+        <activity
+            android:name=".signup.SignUpActivity"
+            android:exported="false">
         </activity>
     </application>
