--- app/src/main/java/com/sopt/now/home/HomeActivity.kt
@@ -1,53 +1,0 @@
-package com.sopt.now.home
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.Fragment
-import com.sopt.now.R
-import com.sopt.now.databinding.ActivityHomeBinding
-import com.sopt.now.mypage.MyPageFragment
-
-class HomeActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityHomeBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityHomeBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        // 액티비티가 처음 생성될 때, HomeFragment를 화면에 표시
-        val currentFragment = supportFragmentManager.findFragmentById(binding.fcvHome.id)
-        if (currentFragment == null) {
-            supportFragmentManager.beginTransaction()
-                .add(binding.fcvHome.id, HomeFragment())
-                .commit()
-        }
-
-        // 바텀 네비게이션 클릭 이벤트 처리 함수 호출
-        clickBottomNavigation()
-    }
-
-    private fun clickBottomNavigation() {
-        // 바텀 네비게이션 아이템 선택 리스너 설정
-        binding.bnvHome.setOnItemSelectedListener{
-            when (it.itemId) {
-                R.id.menu_home-> { // home 선택
-                    replaceFragment(HomeFragment())
-                    true
-                }
-
-                R.id.menu_mypage-> { // my page 선택
-                    replaceFragment(MyPageFragment())
-                    true
-                }
-
-                else -> false
-            }
-        }
-    }
-
-    private fun replaceFragment(fragment: Fragment) {
-        supportFragmentManager.beginTransaction()
-            .replace(binding.fcvHome.id, fragment)
-            .commit()
-    }
-}
