+++ app/src/main/java/com/sopt/now/mypage/MyPageViewModel.kt
@@ -1,11 +1,11 @@
-package com.sopt.now.main
+package com.sopt.now.mypage
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.sopt.now.data.UserData
 
-class MainViewModel : ViewModel() {
+class MyPageViewModel : ViewModel() {
     private val _userInfo = MutableLiveData<UserData>()
     val userInfo: LiveData<UserData> = _userInfo
